jinja_renderer/requirements.txt
black==24.2.0
Django==5.0.2
djangorestframework==3.14.0
Jinja2==3.1.3
django-cors-headers==4.3.1
